+++ examples/navigation/containers/App.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function App({ children }) {
+  return (
+    <div>
+      {children}
+    </div>
+  );
+}