+++ src/EPIC_END.js
@@ -1,0 +1,1 @@
+export const EPIC_END = '@@redux-observable/EPIC_END';