+++ src/index.js
@@ -1,3 +1,4 @@
 export { createEpicMiddleware } from './createEpicMiddleware';
 export { ActionsObservable } from './ActionsObservable';
 export { combineEpics } from './combineEpics';
+export { EPIC_END } from './EPIC_END';