+++ test/typings.ts
@@ -1,7 +1,10 @@
 import { createEpicMiddleware, Epic, combineEpics,
   EpicMiddleware, ActionsObservable } from '../index';
-import { Action } from 'redux';
+import { Action, createStore, applyMiddleware } from 'redux';
 import { Observable } from 'rxjs/Observable';
+import 'rxjs/add/observable/of';
+import 'rxjs/add/operator/mapTo';
+import 'rxjs/add/operator/mergeMap';
 
 const epic1: Epic = (action$, store) =>
   action$.ofType('FIRST')
@@ -12,13 +15,22 @@ const epic1: Epic = (action$, store) =>
 
 const epic2: Epic = (action$, store) =>
   action$.ofType('SECOND')
+    .mapTo('second')
+    .mergeMap(type => Observable.of({ type }));
+
+const epic3: Epic = action$ =>
+  action$.ofType('THIRD')
     .mapTo({
-      type: 'second',
-      payload: store.getState()
+      type: 'third'
     });
 
-const rootEpic1: Epic = combineEpics(epic1, epic2);
-const rootEpic2: Epic = combineEpics(epic1, epic2);
+const epic4: Epic = () =>
+  Observable.of({
+    type: 'fourth'
+  });
+
+const rootEpic1: Epic = combineEpics(epic1, epic2, epic3, epic4);
+const rootEpic2: Epic = combineEpics(epic1, epic2, epic3, epic4);
 
 const epicMiddleware: EpicMiddleware = createEpicMiddleware(rootEpic1);
 
@@ -27,3 +39,12 @@ epicMiddleware.replaceEpic(rootEpic2);
 // should be a constructor that returns an observable
 const actionsSubject = Observable.create(() => {});
 const aoTest: Observable<Action> = new ActionsObservable(actionsSubject);
+
+const reducer = (state = [], action) => state.concat(action);
+const store = createStore(
+  reducer,
+  applyMiddleware(epicMiddleware)
+);
+
+store.dispatch({ type: 'FIRST' });
+store.dispatch({ type: 'SECOND' });